Cleaned Program.cs , moved types to separate files. Changed order in opts so it makes more sense.
