+++ Masarin.IoT.Sensor/Decoders/IMQTTDecoder.cs
@@ -1,0 +1,7 @@
+﻿namespace Masarin.IoT.Sensor
+{
+	interface IMQTTDecoder
+    {
+        void Decode(string timestamp, string device, string topic, byte[] payload);
+    }
+}
