+++ Masarin.IoT.Sensor/Decoders/IMQTTDecoderRegistry.cs
@@ -1,0 +1,7 @@
+﻿namespace Masarin.IoT.Sensor
+{
+	interface IMQTTDecoderRegistry
+    {
+        IMQTTDecoder GetDecoderForNode(string node, string path);
+    }
+}
