+++ Masarin.IoT.Sensor/Decoders/MQTTNullDecoder.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace Masarin.IoT.Sensor
+{
+	class MQTTNullDecoder : MQTTDecoder
+    {
+        public MQTTNullDecoder()
+        {
+        }
+
+        public override void Decode(string timestamp, string device, string topic, byte[] payload)
+        {
+            Console.WriteLine("WARNING: Received data that could not be handled by any decoder ...");
+        }
+    }
+}
