+++ Masarin.IoT.Sensor/Messages/CarMovementMessage.cs
@@ -1,0 +1,10 @@
+﻿namespace Masarin.IoT.Sensor
+{
+	class CarMovementMessage : VehicleMovementMessage
+    {
+        public CarMovementMessage(IoTHubMessageOrigin origin, string timestamp, double heading, double speed) : base(origin, timestamp, heading, speed)
+        {
+            Type = "car";
+        }
+    }
+}
