+++ Masarin.IoT.Sensor/Messages/TelemetryHumidity.cs
@@ -1,0 +1,12 @@
+﻿namespace Masarin.IoT.Sensor
+{
+	class TelemetryHumidity : IoTHubMessage
+    {
+        public int Humidity { get; }
+
+        public TelemetryHumidity(IoTHubMessageOrigin origin, string timestamp, int humidity) : base(origin, timestamp, "telemetry.humidity")
+        {
+            Humidity = humidity;
+        }
+    }
+}
