+++ Masarin.IoT.Sensor/Messages/TelemetryPressure.cs
@@ -1,0 +1,12 @@
+﻿namespace Masarin.IoT.Sensor
+{
+	class TelemetryPressure : IoTHubMessage
+    {
+        public long Pressure { get; }
+
+        public TelemetryPressure(IoTHubMessageOrigin origin, string timestamp, long pressure) : base(origin, timestamp, "telemetry.pressure")
+        {
+            Pressure = pressure;
+        }
+    }
+}
