+++ Masarin.IoT.Sensor/Messages/TelemetrySnowdepth.cs
@@ -1,0 +1,12 @@
+﻿namespace Masarin.IoT.Sensor
+{
+	class TelemetrySnowdepth : IoTHubMessage
+    {
+        public int Snowdepth { get; }
+
+        public TelemetrySnowdepth(IoTHubMessageOrigin origin, string timestamp, int snowdepth) : base(origin, timestamp, "telemetry.Snowdepth")
+        {
+            Snowdepth = snowdepth;
+        }
+    }
+}
