+++ Masarin.IoT.Sensor/Messages/TelemetryWind.cs
@@ -1,0 +1,14 @@
+﻿namespace Masarin.IoT.Sensor
+{
+	class TelemetryWind : IoTHubMessage
+    {
+        public double Speed { get; }
+        public int Direction { get; }
+
+        public TelemetryWind(IoTHubMessageOrigin origin, string timestamp, double speed, int direction) : base(origin, timestamp, "telemetry.wind")
+        {
+            Speed = speed;
+            Direction = direction;
+        }
+    }
+}
